Rename SocialNetwork to Socialnetwork
