+++ apps/socialnetwork/backend/config/routes/health-check.yaml
@@ -1,4 +1,4 @@
 health-check_get:
   path: /health-check
-  controller: Bray\Apps\SocialNetwork\Backend\Controller\HealthCheck\HealthCheckGetController
+  controller: Bray\Apps\Socialnetwork\Backend\Controller\HealthCheck\HealthCheckGetController
   methods:  [GET]
