+++ apps/socialnetwork/backend/src/SocialnetworkBackendKernel.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace Bray\Apps\Socialnetwork\Backend;
+
+use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symfony\Component\HttpKernel\Kernel as BaseKernel;
+use Symfony\Component\Routing\Loader\Configurator\RoutingConfigurator;
+
+class SocialnetworkBackendKernel extends BaseKernel
+{
+    use MicroKernelTrait;
+
+    public function registerBundles(): iterable {
+        $contents = require $this->getProjectDir() . '/config/bundles.php';
+        foreach ($contents as $class => $envs) {
+            if ($envs[$this->environment] ?? $envs['all'] ?? false) {
+                yield new $class();
+            }
+        }
+    }
+
+    public function getProjectDir(): string {
+        return dirname(__DIR__);
+    }
+
+    protected function configureContainer(ContainerConfigurator $container): void {
+        $container->import('../config/{packages}/*.yaml');
+        $container->import('../config/{packages}/' . $this->environment . '/*.yaml');
+
+        if (is_file($this->getProjectDir() . '/config/services.yaml')) {
+            $container->import('../config/services.yaml');
+            $container->import('../config/{services}_' . $this->environment . '.yaml');
+        } elseif (is_file($path = $this->getProjectDir() . '/config/services.php')) {
+            (require $path)($container->withPath($path), $this);
+        }
+    }
+
+    protected function configureRoutes(RoutingConfigurator $routes): void {
+        $routes->import('../config/{routes}/' . $this->environment . '/*.yaml');
+        $routes->import('../config/{routes}/*.yaml');
+
+        if (is_file($this->getProjectDir() . '/config/routes.yaml')) {
+            $routes->import('../config/routes.yaml');
+        } elseif (is_file($path = $this->getProjectDir() . '/config/routes.php')) {
+            (require $path)($routes->withPath($path), $this);
+        }
+    }
+}
