--- src/SocialNetwork/Brays/Application/SearchAll/BraysSearchAllQuery.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Bray\SocialNetwork\Brays\Application\SearchAll;
-
-use Bray\Shared\Domain\Bus\Query\Query;
-
-final class BraysSearchAllQuery implements Query {}
