--- src/SocialNetwork/Brays/Infrastructure/Persistence/Doctrine/BrayDatetime.orm.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<doctrine-mapping xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <embeddable name="Bray\SocialNetwork\Brays\Domain\BrayDatetime">
-        <field name="value" type="string" column="created_at" />
-    </embeddable>
-
-</doctrine-mapping>
