--- src/SocialNetwork/Brays/Infrastructure/Persistence/InMemoryBrayRepository.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Bray\SocialNetwork\Brays\Infrastructure\Persistence;
-
-use Bray\SocialNetwork\Brays\Domain\Contracts\BrayRepository;
-use Bray\SocialNetwork\Brays\Domain\Bray;
-
-final class InMemoryBrayRepository implements BrayRepository
-{
-    private $brays;
-
-    public function __construct() { $this->brays = []; }
-
-    public function searchAll(): array {
-        return $this->brays;
-    }
-
-    public function save(Bray $bray): void {
-        $this->brays[] = $bray;
-    }
-}
