+++ src/Socialnetwork/Brays/Application/Create/CreateBrayCommandHandler.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bray\Socialnetwork\Brays\Application\Create;
+
+use Bray\Shared\Domain\Bus\Command\CommandHandler;
+
+final class CreateBrayCommandHandler implements CommandHandler
+{
+    private BrayCreator $creator;
+
+    public function __construct(BrayCreator $creator) {
+        $this->creator = $creator;
+    }
+
+    public function __invoke(CreateBrayCommand $command): void {
+        ($this->creator)(
+            $command->id(),
+            $command->message(),
+            $command->user(),
+            $command->datetime()
+        );
+    }
+}
