+++ src/Socialnetwork/Brays/Application/SearchAll/BraysSearchAllQuery.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bray\Socialnetwork\Brays\Application\SearchAll;
+
+use Bray\Shared\Domain\Bus\Query\Query;
+
+final class BraysSearchAllQuery implements Query {}
